+++ app/webpage/src/index.html
@@ -3,14 +3,11 @@
   <head>
     <link rel="stylesheet" href="index.css">
     <link rel="shortcut icon" type="image/x-icon" href="favicon.ico">
-    <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
     <script type="module" src="piggo-legends-min.js" defer></script>
     <meta charset="UTF-8">
     <title>Piggo Legends</title>
   </head>
   <body>
-    <div id="root"></div>
-    <canvas id="canvas"/>
+    <div id="root"/>
   </body>
 </html>
